Move RequestPage to NewRequestPage and start request detail
