+++ src/constants/paths.js
@@ -1,7 +1,7 @@
 export const NEW_USER_PATH = '/new-user';
 export const ACCOUNT_PATH = '/account';
 export const LOGIN_PATH = '/login';
-export const REQUEST_PATH = '/request';
+export const REQUEST_PATH = '/requests/:requestId';
 export const REQUEST_SUCCESSFUL_PATH = '/request-successful';
 export const DONATE_PATH = '/donate';
 export const CONTACT_PATH = '/contact';
@@ -9,3 +9,4 @@ export const ABOUT_PATH = '/about';
 export const SEARCH_PATH = '/search';
 export const PRIVACY_POLICY_PATH = '/privacy-policy';
 export const TERMS_OF_SERVICE_PATH = '/terms-of-service';
+export const NEW_REQUEST_PATH = '/new-request';
