+++ src/routes/NewRequest/components/NewRequestPage/index.js
@@ -1,0 +1,3 @@
+import NewRequestPage from './NewRequestPage';
+
+export default NewRequestPage;
