+++ src/routes/Request/components/RequestPage/RequestPage.styles.js
@@ -1,33 +1,16 @@
 export default (theme) => ({
-  buttons: {
-    display: 'flex',
-    justifyContent: 'flex-end',
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(3),
-  },
   paper: {
-    // paddingTop: theme.spacing(3),
-    paddingBottom: theme.spacing(1),
+    padding: theme.spacing(2),
   },
-  optionalDivider: {
-    marginTop: theme.spacing(4),
-    marginBottom: theme.spacing(4),
+  container: {
+    padding: theme.spacing(3),
   },
-  intro: {
+  actionButton: {
+    merginRight: theme.spacing(2),
+    merginLeft: theme.spacing(2),
+  },
+  divider: {
+    marginTop: theme.spacing(2),
     marginBottom: theme.spacing(2),
-  },
-  errorText: {
-    color: 'red',
-    fontWeight: 'bold',
-    display: 'flex',
-    justifyContent: 'flex-end',
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(1),
-  },
-  otherComments: {
-    marginTop: theme.spacing(4),
-  },
-  radio: {
-    marginLeft: theme.spacing(1),
   },
 });
